Split browser parity gates
