+++ docs/DECISIONS.md
@@ -309,3 +309,4 @@
 - TrustedRouter action parsing and tool-argument normalization are separate boundaries. `AgentActionJSONParser` owns extracting a single action object and constructing `AgentAction`; `AgentToolArgumentNormalizer` owns canonical argument keys, tolerated aliases, empty-shell-command repair from explicit nearby prose, and minimum-argument policy. Keep provider transport, JSON-object scanning, prose recovery, and tool schema alias policy in separate files so model-output robustness can improve without turning the parser into a tool registry.
 - Created-thread insertion should go through `WorkspaceModel.insertCreatedThread` unless there is a documented reason not to. New chat, fork, compact, duplicate, worktree-open, and automation-run flows should share selection clearing, terminal sync, project touch, persistence, and top-bar refresh semantics so sidebar bulk-selection and project state do not drift by creation path.
 - Browser workflow orchestration lives in `WorkspaceBrowserWorkflow`, not inline in `WorkspaceModel`. The workflow owns URL resolution calls, invalid-address copy, history/navigation delegation, snapshot fetch begin/complete semantics, stale-fetch protection, and browser-comment dispatch. `WorkspaceModel` owns actor-bound state, async page fetching, and top-bar refresh after workflow results.
+- Browser parity gates live in `ParityBrowserGateTests`, not the broad `ParityGateTests` catch-all. Browser surface ownership, snapshot extraction, workflow state transitions, location resolving, browser integration-test ownership, and HTML browser-rendering boundaries should stay together so active browser work does not keep inflating the general architecture-gate file.
